+++ NeuralNet/MathHelper.cs
@@ -8,14 +8,14 @@ namespace NeuralNet {
 
     public static class MathHelper {
 
-        public static float Normalize(float input, float min, float max) {
+        public static double Normalize(double input, double min, double max) {
             var val = (input - min) / (max - min);
             return val;
         }
 
-        public static IEnumerable<float> Normalize(IEnumerable<float> collection) {
-            float min = collection.Min();
-            float max = collection.Max();
+        public static IEnumerable<double> Normalize(IEnumerable<double> collection) {
+            double min = collection.Min();
+            double max = collection.Max();
 
             return collection.Select(f => Normalize(f, min, max));
         }
